EXECUTOR WORKS on best-case input scenario, a lot of things still to do
