+++ infile.txt
@@ -1,7 +1,7 @@
 ABC
+Zuzka
 abcd
 Anka
-Zuzka
 123
 def
 AAAA
