+++ outfile.txt
@@ -1,3 +1,5 @@
-abcd
-Anka
-Zuzka
+builtins.c
+check_input.c
+evaluator.c
+executor.c
+exit.c
